refactor states namespace
