+++ src/States/AppState.cs
@@ -1,0 +1,9 @@
+﻿namespace EnvironmentControl.States {
+    public class AppState {
+        public AppState(EditStatus editStatus) {
+            EditStatus = editStatus;
+        }
+
+        public EditStatus EditStatus { get; }
+    }
+}
