+++ src/States/EditStatus.cs
@@ -1,0 +1,6 @@
+﻿namespace EnvironmentControl.States {
+    public enum EditStatus {
+        Normal, 
+        Editing
+    }
+}
