--- src/ViewModels/AppState.cs
@@ -1,9 +1,0 @@
-﻿namespace EnvironmentControl.ViewModels {
-    public class AppState {
-        public AppState(State state) {
-            State = state;
-        }
-
-        public State State { get; }
-    }
-}
