--- src/ViewModels/State.cs
@@ -1,6 +1,0 @@
-﻿namespace EnvironmentControl.ViewModels {
-    public enum State {
-        Normal, 
-        Editing
-    }
-}
