+++ src/Views/MainWindow.xaml.cs
@@ -2,6 +2,7 @@
 using System.Windows;
 using EnvironmentControl.ViewModels;
 using System.Windows.Input;
+using EnvironmentControl.States;
 
 namespace EnvironmentControl.Views {
     public partial class MainWindow {
